+++ eStoreManager-SERVER/src/main/java/com/example/store/payload/product_management/request/CreateProductRequest.java
@@ -24,7 +24,7 @@ public class CreateProductRequest {
     private String barcode;
 
     @Min(0)
-    private Long quantities;
+    private Float quantities;
 
     public String getName() {
         return name;
@@ -42,7 +42,7 @@ public class CreateProductRequest {
         return barcode;
     }
     
-    public Long getQuantities() {
+    public Float getQuantities() {
         return quantities;
     }
